# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BottleCRM API",
    "version": "1.0.0",
    "description": "Multi-tenant CRM API with JWT authentication",
    "pluginName": "BottleCRM智能助手",
    "pluginDesc": "BottleCRM客户关系管理系统AI助手，支持查询客户数据、创建线索任务、分析销售业绩等功能",
    "contact": "support@bottlecrm.com",
    "x-huawei-plugin-name": "BottleCRM智能助手",
    "x-huawei-plugin-desc": "BottleCRM客户关系管理系统AI助手，支持查询客户数据、创建线索任务、分析销售业绩等功能",
    "x-huawei-version": "1.0.0",
    "x-huawei-contact": "support@bottlecrm.com"
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/contacts": {
      "get": {
        "summary": "Get all contacts for organization",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of contacts"
          }
        }
      },
      "post": {
        "summary": "Create a new contact",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName"
                ],
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "department": {
                    "type": "string"
                  },
                  "street": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "postalCode": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "accountId": {
                    "type": "string",
                    "description": "UUID of the account to associate with this contact"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "summary": "Get a specific contact by ID",
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Contact ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact details"
          },
          "404": {
            "description": "Contact not found"
          }
        }
      }
    },
...
  },
...
}